refactor: fix url inconsistency
