+++ packages/client/src/main.tsx
@@ -4,10 +4,10 @@ import { createBrowserRouter, RouterProvider } from "react-router-dom";
 
 // Intern
 import { Navigation } from "$components/Navigation";
-import { LoginPage } from "$pages/auth/LoginPage";
+import { MainPage } from "$pages/MainPage";
 import { HomePage } from "$pages/HomePage";
+import { LoginPage } from "$pages/auth/LoginPage";
 import { RegisterPage } from "$pages/auth/RegisterPage";
-import { MainPage } from "$pages/general/MainPage";
 import { WorkoutsPage } from "$pages/workout/WorkoutsOverview";
 import "./index.css";
 
@@ -16,11 +16,7 @@ const root = document.getElementById("root")!;
 const router = createBrowserRouter([
   {
     path: "/",
-    element: (
-      <>
-        <HomePage/>
-      </>
-    )
+    element: <MainPage/>
   },
   {
     path: "/login",
@@ -34,7 +30,7 @@ const router = createBrowserRouter([
     path: "/home",
     element: (
       <>
-        <MainPage/>
+        <HomePage/>
         <Navigation/>
       </>
     )
