+++ packages/client/src/pages/HomePage.tsx
@@ -1,22 +1,23 @@
 import React from "react";
-import { Link } from "react-router-dom";
+
+// Intern
+import { BackgroundLayout } from "$components/Background";
+import { Navigation } from "$components/Navigation";
 
 export const HomePage: React.FC = () => {
   return (
-    <div className="flex flex-col items-center justify-center h-screen bg-gray-100">
-      <h1 className="text-4xl font-bold text-blue-600 mb-6">Welcome to Momentum!</h1>
-      <div className="space-x-4">
-        <Link to="/login">
-          <button className="bg-blue-500 text-white px-4 py-2 rounded hover:bg-blue-700">
-            Login
-          </button>
-        </Link>
-        <Link to="/register">
-          <button className="bg-green-500 text-white px-4 py-2 rounded hover:bg-green-700">
-            Register
-          </button>
-        </Link>
+    <BackgroundLayout>
+      <div className="flex flex-col items-center space-y-4 w-full">
+        <h1 className="text-2xl font-bold text-black">Momentum</h1>
+        <p className="text-lg text-black">Welcome back, <span className="font-bold">John.</span></p>
+        <button 
+          className="w-full bg-green-500 hover:bg-green-600 text-white font-bold py-2 px-4 rounded"
+          onClick={() => alert("Starting workout...")}
+        >
+          Start Workout: Push
+        </button>
       </div>
-    </div>
+      <Navigation/>
+    </BackgroundLayout>
   );
 };
